refactor: 이벤트 디렉토리 위치 이동 - #308
