+++ product/src/main/java/com/back/product/app/facade/ProductFacade.java
@@ -9,9 +9,9 @@ import com.back.product.dto.command.*;
 import com.back.product.dto.model.*;
 import com.back.product.dto.request.*;
 import com.back.product.dto.response.*;
-import com.back.product.dto.event.spring.ProductCreationCompletedEvent;
-import com.back.product.dto.event.spring.ProductDeletionCompletedEvent;
-import com.back.product.dto.event.spring.ProductUpdateCompletedEvent;
+import com.back.product.event.spring.ProductCreationCompletedEvent;
+import com.back.product.event.spring.ProductDeletionCompletedEvent;
+import com.back.product.event.spring.ProductUpdateCompletedEvent;
 import com.back.product.mapper.*;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
