+++ product/src/main/java/com/back/product/event/spring/ProductUpdateCompletedEvent.java
@@ -1,6 +1,7 @@
-package com.back.product.dto.event.kafka;
+package com.back.product.event.spring;
 
-import com.back.common.event.KafkaPayload;
+import com.back.product.dto.model.OptionDto;
+import com.back.product.dto.model.ProductInfoDto;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotEmpty;
@@ -11,17 +12,17 @@ import org.hibernate.validator.constraints.URL;
 import java.util.List;
 
 @Builder
-public record ProductCreatedPayload(
-        @NotNull(message = "상품 정보 페이로드는 필수입니다.")
+public record ProductUpdateCompletedEvent(
+        @NotNull(message = "상품 정보 DTO는 필수입니다.")
         @Valid
-        ProductInfoPayload productInfo,
+        ProductInfoDto productInfoDto,
 
-        @NotEmpty(message = "옵션 페이로드 목록은 비어 있을 수 없습니다.")
+        @NotEmpty(message = "옵션 DTO 목록은 비어 있을 수 없습니다.")
         @Valid
-        List<OptionPayload> options,
+        List<OptionDto> optionDtos,
 
         @NotBlank(message = "썸네일 URL은 필수입니다.")
         @URL(message = "유효한 URL 형식이 아닙니다.")
         String thumbnailUrl
-) implements KafkaPayload {
+) {
 }
