+++ product/src/test/java/com/back/product/app/ProductSyncEventPublishTest.java
@@ -2,19 +2,18 @@ package com.back.product.app;
 
 import com.back.product.adapter.out.event.ProductKafkaEventPublisher;
 import com.back.product.app.usecase.command.ProductDocumentUseCase;
-import com.back.product.dto.event.kafka.ProductCreatedPayload;
-import com.back.product.dto.event.kafka.ProductDeletedPayload;
-import com.back.product.dto.event.kafka.ProductInfoPayload;
-import com.back.product.dto.event.kafka.ProductUpdatedPayload;
+import com.back.product.event.kafka.ProductCreatedPayload;
+import com.back.product.event.kafka.ProductDeletedPayload;
+import com.back.product.event.kafka.ProductInfoPayload;
+import com.back.product.event.kafka.ProductUpdatedPayload;
 import com.back.product.dto.model.BrandDto;
 import com.back.product.dto.model.CategoryDto;
 import com.back.product.dto.model.OptionDto;
 import com.back.product.dto.model.ProductInfoDto;
-import com.back.product.dto.event.spring.ProductCreationCompletedEvent;
-import com.back.product.dto.event.spring.ProductUpdateCompletedEvent;
-import com.back.product.dto.event.spring.ProductDeletionCompletedEvent;
+import com.back.product.event.spring.ProductCreationCompletedEvent;
+import com.back.product.event.spring.ProductUpdateCompletedEvent;
+import com.back.product.event.spring.ProductDeletionCompletedEvent;
 import jakarta.validation.ConstraintViolationException;
-import jakarta.validation.Validator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
